Cleanup unused code.
